+++ slack-app-backend/src/Configuration/AppConfiguration.cs
@@ -7,5 +7,6 @@ namespace SlackAppBackend.Configuration
     {
         public Uri SystemEventServiceUri => new Uri(Environment.GetEnvironmentVariable("SYSTEM_EVENTS_SVC_URI"));
         public string SlackBotUserOAuthAccessToken => Environment.GetEnvironmentVariable("SLACK_OAUTHACCESS_TOKEN");
+        public string SlackSignedSecret => Environment.GetEnvironmentVariable("SLACK_SIGNED_SECRET");
     }
 }